create answers app move answer and related functions to answer application
